### 2023-09-02 #606 @HaoZeke
`fastMatMR`: Fast Matrix Market I/O
Submitting Author Name: Rohit Goswami
Submitting Author Github Handle: @HaoZeke
Repository:  https://github.com/HaoZeke/fastMatMR
Version submitted:
Submission type: Standard

Version accepted: TBD

---

```
Package: fastMatMR
Title: "fastMatMR: High-Performance Matrix Market File Operations in R"
Version: 1.0.0.0
    person("Rohit", "Goswami", , "rgoswami@ieee.org", role = c("aut", "cre"),
Description: "fastMatMR is an R package offering high-performance read and write operations for Matrix Market files. It acts as a thin wrapper around the 'fast_matrix_market' C++ library, offering speed and extended support for Matrix Market formats. Unlike other R packages, fastMatMR supports not just sparse matrices but also dense vectors and matrices. This makes it a versatile choice for dealing with .mtx files in R."
SystemRequirements: C++17
    cpp11
    Matrix,
    microbenchmark,
```

The [matrix market exchange formats](https://math.nist.gov/MatrixMarket/formats.html) are crucial to much of the ecosystem. The fastMatMR package focuses on high-performance read and write operations for Matrix Market files, serving as a key tool for data extraction in computational and data science pipelines.

Data scientists who might want to load and test the NIST matrices which include:

> comparative studies of algorithms for numerical linear algebra, featuring nearly 500 sparse matrices from a variety of applications, as well as matrix generation tools and services.

Additionally, this makes its simpler to interface to `scipy` and the rest of the data science ecosystem. This also includes working with the [Tensor Algebra Compiler](http://tensor-compiler.org/docs/tensors.html) (TACO).

The `Matrix` package in `R` can perform similar operations but only for sparse matrices. The `fastMatMR` package not only provides enhanced performance but also extends support to dense matrices and vectors in `base R`, thus offering a more versatile solution.

We have both [read](https://haozeke.github.io/fastMatMR/articles/fmm_read_bench.html) and [write](https://haozeke.github.io/fastMatMR/articles/fmm_write_bench.html) performance vignettes backing up the claims made.

The package passes `pkcheck`: https://github.com/HaoZeke/fastMatMR/issues/18, though the review bot disagrees :)
