### 2023-09-02 #606 @HaoZeke
Also, not part of the issue template, but the package passes `pkcheck`: https://github.com/HaoZeke/fastMatMR/issues/18
