### 2023-09-02 fe999d7
MAINT: Fix a typo, rerender benchmarks
