### 2023-09-02 #606 @HaoZeke
FWIW there's `.covignore` which is not being picked up by the bot here (works on the repo, https://github.com/HaoZeke/fastMatMR/issues/18). Package coverage is:

```R
> covr::package_coverage()
fastMatMR Coverage: 94.69%
src/from_file.cpp: 93.88%
src/to_file.cpp: 94.23%
R/fastMatMR-package.R: 100.00%
```
