### 2023-09-05 #606 @mpadge
@HaoZeke is indeed right there - the failing coverage check is [a bug on our side](https://github.com/ropensci-review-tools/pkgcheck-action/issues/43), and not this submission.
