### 2023-09-12 #606 @maelle
:wave: @HaoZeke! Nice to see you as an author this time around, thank you for your submission! :grinning:

A few comments/questions before we proceed to the reviewer search:

- Why add the bundled files to .covrignore, if they contribute to the functionality of the package?
- I see you use conventional commits, nice. How is the changelog generated? I see https://github.com/HaoZeke/fastMatMR/issues/17 Disclaimer: I contribute to https://github.com/cynkra/fledge/
- In the example where you use `->` I'd suggest using `<-` instead since the former is more rare.
- In the README you write "writing and reading from other R objects". Could you please elaborate, adding a few sentences about this to the README?
- To me the second part of the README, "Development", belongs to the contributing guide, so you could move the content and instead point contributors to the contributing guide.
- In the contributing guide could you please add a link to pixi? I for instance do not know what it is.
- In the issue tracker there are a few enhancement issues. What's the timeline for them, are they "nice to have" or do you intend to tackle them soon? If so should the review wait?

Thank you!
