### 2023-09-12 fe7a5d4
DOC: Refactor and add vignette links
