### 2023-09-14 #606 @maelle
> This is because there is no need to expose bindings of the inner workings of the fast_matrix_market C++ library to R users. This approach is also taken by other libraries which bundle C++ libraries, like [RCppEigen](https://github.com/RcppCore/RcppEigen/blob/master/.covrignore?rgh-link-date=2023-09-12T11%3A37%3A43Z).

To clarify, the question was about tests specifically. Part of the code in inst/ is covered by tests anyway, so why not include them (and remove the unused code)?

We do not have strict requirements on testing coverage for bundled code yet, but we can improve our guidance based on cases like this submission, so this discussion is important.
