### 2023-09-14 #606 @HaoZeke
> > This is because there is no need to expose bindings of the inner workings of the fast_matrix_market C++ library to R users. This approach is also taken by other libraries which bundle C++ libraries, like [RCppEigen](https://github.com/RcppCore/RcppEigen/blob/master/.covrignore?rgh-link-date=2023-09-12T11%3A37%3A43Z).
>
> To clarify, the question was about tests specifically. Part of the code in inst/ is covered by tests anyway, so why not include them (and remove the unused code)?

Mostly maintainence. I have actually stripped the (already small) library of header files which are not used at all (Eigen / Armadillo / other C++ library compatibility files).

In general, for bundled libraries it is a good idea to *not* tamper with the upstream source as much as possible (it makes it easier to update to new versions).

> We do not have strict requirements on testing coverage for bundled code yet, but we can improve our guidance based on cases like this submission, so this discussion is important.

Yup, I understand. Policy wise I'd be even stricter, and say that bundled libraries must not be touched at all (other than whole-file removals if they are not relevant and self contained).

This is standard practice in bundling libraries in C++ too, where projects might have different linting rules and `external` files are ignored instead of refactored in-place within the project. The idea is that way upstream developers can weigh in more easily on the code.
