### 2023-09-15 f369dea
DOC: Update with rOpenSci badge
