### 2023-09-22 #606 @osorensen
@HaoZeke and @maelle, here is my initial review. My comments at the bottom explain a bit more about the boxes I haven't checked.

## Package Review

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☐ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☐ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☐ **Function Documentation:** for all exported functions
-   ☐ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☐ **Functionality:** Any functional claims of the software been confirmed.
-   ☐ **Performance:** Any performance claims of the software been confirmed.
-   ☒ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

-   ☐ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review comments

#### Dense matrices with the matrix package

In README, you state that the `Matrix` package only handles sparse objects. This is not correct.

    library(Matrix)

The command below lists all dense matrix types supported by the `Matrix` package.

    showClass("denseMatrix")

    ## Virtual Class "denseMatrix" [package "Matrix"]
    ##
    ## Slots:
    ##
    ## Class:  integer     list
    ##
    ## Class "replValueSp", by class "Matrix", distance 2
    ##
    ## Class "pcorMatrix", by class "dppMatrix", distance 4

Below is an example of a dense matrix from the `Matrix` package:

    (dense_matrix <- as(matrix(1:10, ncol = 2), "dMatrix"))

    ## [5,]    5   10

    str(dense_matrix)

    ##   ..@ factors : list()

#### Statement of need

The README file contains a section titled “Why?”, but it does not state explicitly what problems the software is designed to solved nor what the target audience is.

#### Vignettes

There are vignettes, but they are very brief. Please provide some more extended examples, with more text explaining what is being done.

#### Examples

I understand that the `fmm_to_*` functions need an mtx file to work, and I guess it’s therefore the examples with these functions are marked with `\dontrun{}`. However, wouldn’t it possible to define a character vector in R that contains the required input, and then feed this to `fmm_to_*` in order to have an example that can actually be run?
