### 2023-09-22 #606 @HaoZeke
Just a quick clarification, @osorensen, the `readme` wording is a bit unclear but it states in the previous section that the goal is the reading and writing of `Matrix Market` files. For that, it is true that `Matrix` does not support dense matrices:

```R
> library("Matrix")
> (dense_matrix <- as(matrix(1:10, ncol = 2), "dMatrix"))
> writeMM(dense_matrix, "mat.mtx")
Error in (function (classes, fdef, mtable)  :
  unable to find an inherited method for function ‘writeMM’ for signature ‘"dgeMatrix"’
```

The longer explanation is part of the ROpenSci issue (intended audience, use case) but I will add it to the readme soon.
