### 2023-10-07 #606 @czeildi
---

All in all, I could easily install the package, understand its functionality
and use it in some of my (simple) own examples without having previous knowledge of matrix market. I could navigate the code and mostly understand it (I am not experienced with cpp11 code).

I have a couple of comments, detailed below. Most of them are non-blocking and
might be somewhat opinionated so feel free to apply them as you see fit (besides the ones where I did not tick the corresponding box)

- Performance: I did not re-run the benchmark vignettes myself but tried the package
  out with some moderately large matrices and it felt reasonably fast. 👍🏼
- it seems the lib cannot handle paths with `~` (?) If this cannot be made to work easily, some error message would be very helpful, I think it would be needed
  to be added here: https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L26 The corresponding read file returns an error message: 'cannot open file' which is more helpful.
  (I am guessing the issue is related to `~` in path because `write_fmm(vec, "~/vector.mtx")` returned false for me while `write_fmm(vec, "/home/ildi_home/vector.mtx")` worked)
- I really appreciated the precise error message in
```r
vec_to_fmm(1L, "vector.mtx")
#> Error: Invalid input type, expected 'double' actual 'integer'
```
  However, I was wondering if it could actually accept integers as well?

- improve error message? It is a matrix, just sparse, not plain
```r
mat_to_fmm(sparse_mat, "sparse_matrix.mtx")
#> Error in mat_to_fmm(sparse_mat, "sparse_matrix.mtx") : object is not a matrix
```

- improve error message?
```r
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
sparse_to_fmm(mat, "matrix.mtx")
#> Error: Invalid input type, expected 'double' actual 'NULL'
```
- future suggestion: support reading `.mtx.gz` files as I can download from for example https://math.nist.gov/MatrixMarket/data/NEP/quebec/qh1484.html ?
- consider increased consistency of function names? e.g `fmm_to_sparseMatrix` vs `sparse_to_fmm`. `sparse_to_fmm` was a bit confusing to me in that if I understand correctly, the package is 'fast', not the target format itself (?). matrix and vector is abbreviated to spare a few characters, while `sparse_matrix` is not. A vague suggestion: possibly follow the pattern of `fmm_write_vector`, `fmm_read_to_vector` for all formats?

- I did not tick function documentation because although they exist, argument names
  (`fname` vs `filename`) are not always consistent between the documentation and the code
  itself
- I did not tick Examples, because although the examples run, but only in the right order, i.e. if I first run the examples of the write
  functions then the examples of the read functions run correctly but they are not self contained
- readme: "Unlike the Matrix package" : maybe rephrase? it is a bit off-putting for me to start the whole readme with a negative comparison, although
  I understand that a core part of the motivation for this package is that Matrix does not support everything and not fast enough. Even changing the order of the two parts of the sentence would feel better for me but this is really personal
- In the readme: "we support / the package supports". Consider unifying these two wording styles
- rename `basic_usage` to getting started? The vignette title is getting started but you can see `basic_usage` in the url
- clarify what the `addendum` is about in the getting started readme? If I understand correctly, it demonstrates the incorrect handling of `NA` by `Matrix` package?
- Maybe add a bit more context on why would I want to read back in `Python`? Possibly rather focus on being a language agnostic format?
- consider moving the vignette summary to the beginning/remove it? It felt to me like repeating basically the same information as the beginning of the vignette
- Did you try out logarithmic scale for the benchmark plots? Would they possibly be more readable?
- In several function documentation you write "This function has no return value." This is technically not true, all functions return `NULL` if they have no other return value
  but rather these functions are not called for their return value, but for their side effect? Maybe see how other packages handle write functions?
- `write_fmm` example: `sparse_mat` variable is overwritten before it is written to disk so it is not easy to actually run both of the two variants included
- readme sentence: "Similarly, we support writing and reading from other R objects (e.g. standard R vectors and matrices), as seen in the basic use vignette."
  Consider rephrasing a bit, e.g. refer to 'getting started' vignette instead of basic use, possibly as described in the vignette / as you can see in the vignette instead of as seen?
- I am not sure how, but I think there is a relatively easy way to have the getting started vignette show up as a separate menu item in the packagedown website and not just among articles, that could improve its discoverability

- I see that R CMD check seem to give a note about pixi files, it seems `pixi.*` might not
  be the correct syntax in `.Rbuildignore`? (I do not know what the correct syntax would be,
  but the check gives a note both in github actions and locally it seems. If all else fails, adding the files one by one instead of a wildcard syntax definitely works)
- package version in `DESCRIPTION` does not match package version in `news`, is this intended?
- Ideally, tests would not write to the project root in my opinion even if those files are added
  to `.gitignore`, but rather write to a temp directory. https://testthat.r-lib.org/articles/test-fixtures.html might help find a more robust solution.
- the readability of read unit tests could be improved if all tests are self contained and independent of each other: e.g. you can run any `test_that` block in itself and it is a working test (no setup code before `test_that` blocks ideally)
- I think there is no need to add `library(testthat)` to test files
- consistency of indentation styles could be a bit improved, consider running `styler::style_pkg`, see for example https://github.com/HaoZeke/fastMatMR/blob/main/tests/testthat/test-write_fmm.R#L34
- this is really overly pedantic, but you could consider specifying which rule to ignore in the `#nolint` comments instead of ignoring every rule (it is considered generally a best practice, but does not have real impact here). This is not needed if you would need to ignore several rules
- pre commit github workflow name is a bit misleading to me, I understand you could add the same functionality as a pre commit hook, but a github action is a post commit/push thing, maybe title it after what it does, e.g. style checking?
- where is this included from? I only see `cpp11.cpp` https://github.com/HaoZeke/fastMatMR/blob/main/src/from_file.cpp#L10
  My IDE underlined this with red, but it might be fine, I am not familiar with cpp bindings in R packages
- commented out code should be removed (https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L20)
- what happens if cpp lib call fails for some reason? Will the file handle be closed properly? See https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L30
- possibly move rebuild-benchmarks to a tools sub directory or similar? https://r-pkgs.org/misc.html#sec-misc-tools.
