### 2023-10-22 HaoZeke/fastMatMR#24 @HaoZeke
ENH: Handle `~` or be more explicit
> it seems the lib cannot handle paths with ~ (?) If this cannot be made to work easily, some error message would be very helpful, I think it would be needed
to be added here: https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L26 The corresponding read file returns an error message: 'cannot open file' which is more helpful.
(I am guessing the issue is related to ~ in path because write_fmm(vec, "~/vector.mtx") returned false for me while write_fmm(vec, "/home/ildi_home/vector.mtx") worked)

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.

Generally, since `R` needs to work on Windows as well, supporting `~` will be hard, but a better error message would be nicer.
