### 2023-10-22 HaoZeke/fastMatMR#25 @HaoZeke
ENH: Handle integers in input files
> vec_to_fmm(1L, "vector.mtx")
#> Error: Invalid input type, expected 'double' actual 'integer'

> However, I was wondering if it could actually accept integers as well?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
