### 2023-10-22 HaoZeke/fastMatMR#26 @HaoZeke
MAINT: Better error messages for sparse vs non-sparse matrics
> improve error message? It is a matrix, just sparse, not plain
```
mat_to_fmm(sparse_mat, "sparse_matrix.mtx")
#> Error in mat_to_fmm(sparse_mat, "sparse_matrix.mtx") : object is not a matrix
```
First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
