### 2023-10-22 HaoZeke/fastMatMR#27 @HaoZeke
MAINT: Better error messages for `sparse_to_fmm`
> improve error message?
```
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
sparse_to_fmm(mat, "matrix.mtx")
#> Error: Invalid input type, expected 'double' actual 'NULL'
```

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.

The wrong function is called but the error message is confusing.
