### 2023-10-22 HaoZeke/fastMatMR#28 @HaoZeke
ENH: Handle `.mtx.gz`
> future suggestion: support reading .mtx.gz files as I can download from for example https://math.nist.gov/MatrixMarket/data/NEP/quebec/qh1484.html ?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
