### 2023-10-22 HaoZeke/fastMatMR#29 @HaoZeke
ENH: Better named functions
> consider increased consistency of function names? e.g fmm_to_sparseMatrix vs sparse_to_fmm. sparse_to_fmm was a bit confusing to me in that if I understand correctly, the package is 'fast', not the target format itself (?). matrix and vector is abbreviated to spare a few characters, while sparse_matrix is not. A vague suggestion: possibly follow the pattern of fmm_write_vector, fmm_read_to_vector for all formats?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.

This is a breaking change, so I'm not too sure, but then again, better now than any other time. I'm OK with `fmm` myself, since its the package name, but yeah consistency is a good target.
