### 2023-10-22 HaoZeke/fastMatMR#30 @HaoZeke
DOC: Consistent argument names
> I did not tick function documentation because although they exist, argument names
(`fname` vs `filename`) are not always consistent between the documentation and the code
itself

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
