### 2023-10-22 HaoZeke/fastMatMR#31 @HaoZeke
DOC,MAINT: Idempotent examples
> I did not tick Examples, because although the examples run, but only in the right order, i.e. if I first run the examples of the write
functions then the examples of the read functions run correctly but they are not self contained

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
