### 2023-10-22 HaoZeke/fastMatMR#32 @HaoZeke
DOC: Reword `readme`
> readme: "Unlike the Matrix package" : maybe rephrase? it is a bit off-putting for me to start the whole readme with a negative comparison, although
I understand that a core part of the motivation for this package is that Matrix does not support everything and not fast enough. Even changing the order of the two parts of the sentence would feel better for me but this is really personal

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
