### 2023-10-22 HaoZeke/fastMatMR#33 @HaoZeke
DOC: `readme` consistency (style)
> In the readme: "we support / the package supports". Consider unifying these two wording styles

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
