### 2023-10-22 HaoZeke/fastMatMR#34 @HaoZeke
DOC: Rename vignettes
> rename basic_usage to getting started? The vignette title is getting started but you can see basic_usage in the url

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
