### 2023-10-22 HaoZeke/fastMatMR#35 @HaoZeke
DOC: Move addenum
> clarify what the addendum is about in the getting started readme? If I understand correctly, it demonstrates the incorrect handling of NA by Matrix package?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
