### 2023-10-22 HaoZeke/fastMatMR#36 @HaoZeke
DOC: Add details on language interop
> Maybe add a bit more context on why would I want to read back in Python? Possibly rather focus on being a language agnostic format?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
