### 2023-10-22 HaoZeke/fastMatMR#37 @HaoZeke
DOC: Vignette reduction (getting started)
> consider moving the vignette summary to the beginning/remove it? It felt to me like repeating basically the same information as the beginning of the vignette

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
