### 2023-10-22 HaoZeke/fastMatMR#38 @HaoZeke
DOC: Plot scales
> Did you try out logarithmic scale for the benchmark plots? Would they possibly be more readable?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
