### 2023-10-22 HaoZeke/fastMatMR#39 @HaoZeke
DOC: Function help details are incorrect
> In several function documentation you write "This function has no return value." This is technically not true, all functions return NULL if they have no other return value but rather these functions are not called for their return value, but for their side effect? Maybe see how other packages handle write functions?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
