### 2023-10-22 HaoZeke/fastMatMR#40 @HaoZeke
DOC: Vignette variables are overwritten
> write_fmm example: sparse_mat variable is overwritten before it is written to disk so it is not easy to actually run both of the two variants included

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
