### 2023-10-22 HaoZeke/fastMatMR#41 @HaoZeke
MAINT: Vignette accessibility
> I am not sure how, but I think there is a relatively easy way to have the getting started vignette show up as a separate menu item in the packagedown website and not just among articles, that could improve its discoverability

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
