### 2023-10-22 HaoZeke/fastMatMR#42 @HaoZeke
DOC: Make `fmm_to*` runnable
> I understand that the fmm_to_* functions need an mtx file to work, and I guess it’s therefore the examples with these functions are marked with \dontrun{}. However, wouldn’t it possible to define a character vector in R that contains the required input, and then feed this to fmm_to_* in order to have an example that can actually be run?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1731654678
