### 2023-10-22 HaoZeke/fastMatMR#43 @HaoZeke
MAINT: Get `R CMD check` to give less notes
> I see that R CMD check seem to give a note about pixi files, it seems pixi.* might not
be the correct syntax in .Rbuildignore? (I do not know what the correct syntax would be,
but the check gives a note both in github actions and locally it seems. If all else fails, adding the files one by one instead of a wildcard syntax definitely works)

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
