### 2023-10-22 HaoZeke/fastMatMR#44 @HaoZeke
TST: Do not write to project root
> Ideally, tests would not write to the project root in my opinion even if those files are added
to .gitignore, but rather write to a temp directory. https://testthat.r-lib.org/articles/test-fixtures.html might help find a more robust solution.

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
