### 2023-10-22 HaoZeke/fastMatMR#45 @HaoZeke
TST: Refactor unit tests
> the readability of read unit tests could be improved if all tests are self contained and independent of each other: e.g. you can run any test_that block in itself and it is a working test (no setup code before test_that blocks ideally)

> I think there is no need to add library(testthat) to test files

> consistency of indentation styles could be a bit improved, consider running styler::style_pkg, see for example https://github.com/HaoZeke/fastMatMR/blob/main/tests/testthat/test-write_fmm.R#L34

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
