### 2023-10-22 HaoZeke/fastMatMR#46 @HaoZeke
MAINT: Misc review comments
> where is this included from? I only see cpp11.cpp https://github.com/HaoZeke/fastMatMR/blob/main/src/from_file.cpp#L10
My IDE underlined this with red, but it might be fine, I am not familiar with cpp bindings in R packages

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
