### 2023-10-22 HaoZeke/fastMatMR#47 @HaoZeke
CI: Rename workflow
> pre commit github workflow name is a bit misleading to me, I understand you could add the same functionality as a pre commit hook, but a github action is a post commit/push thing, maybe title it after what it does, e.g. style checking?

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
