### 2023-10-22 HaoZeke/fastMatMR#48 @HaoZeke
MAINT: Remove commented code
> commented out code should be removed (https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L20)

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
