### 2023-10-22 HaoZeke/fastMatMR#49 @HaoZeke
ENH: Better error handling
> what happens if cpp lib call fails for some reason? Will the file handle be closed properly? See https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L30

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.

I'm not sure this can even happen though..
