### 2023-10-22 HaoZeke/fastMatMR#50 @HaoZeke
MAINT: Refactor build folders
> possibly move rebuild-benchmarks to a tools sub directory or similar? https://r-pkgs.org/misc.html#sec-misc-tools.

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
