### 2023-10-22 HaoZeke/fastMatMR#51 @HaoZeke
MAINT:  More scoped `nolint` rules
> this is really overly pedantic, but you could consider specifying which rule to ignore in the #nolint comments instead of ignoring every rule (it is considered generally a best practice, but does not have real impact here). This is not needed if you would need to ignore several rules

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1751660212.
