### 2023-10-22 8c4bf9e
MAINT: Cleanup basic usage vignette
Closes https://github.com/HaoZeke/fastMatMR/issues/37
Closes https://github.com/HaoZeke/fastMatMR/issues/34
