### 2023-10-22 a76038a
DOC: Rewrite README
Closes #35. Closes #36. Closes #33. Closes #32.
