### 2023-10-22 HaoZeke/fastMatMR#42 @HaoZeke
I don't think this is actually possible, since we call C++ functions on the actual files instead of operating on strings from `R`..
