### 2023-10-22 HaoZeke/fastMatMR#31 @HaoZeke
Also not much to do, the read and write functions need files or data to operate on...
