### 2023-10-22 e9cc389
TST: Add tempdir fixtures and refactor tests
Closes #44. Closes #45.
