### 2023-10-22 3a53dbb
MAINT: Rework parameters for consistency
fname -> filename

Closes gh-30.
