### 2023-10-22 65d192e
ENH: Expand paths before passing to C++
Closes gh-24.
