### 2023-10-22 d22d26e
ENH: Expand paths for fmm_to_ functions
Also part of gh-24.
