### 2023-10-22 6c2b7c3
MAINT: Add the auto-gen files for gh-24
