+++ R/cpp11.R
@@ -1,15 +1,15 @@
 # Generated by cpp11: do not edit by hand
 
-fmm_to_vec <- function(filename) {
-  .Call(`_fastMatMR_fmm_to_vec`, filename)
+cpp_fmm_to_vec <- function(filename) {
+  .Call(`_fastMatMR_cpp_fmm_to_vec`, filename)
 }
 
-fmm_to_mat <- function(filename) {
-  .Call(`_fastMatMR_fmm_to_mat`, filename)
+cpp_fmm_to_mat <- function(filename) {
+  .Call(`_fastMatMR_cpp_fmm_to_mat`, filename)
 }
 
-fmm_to_sparse_Matrix <- function(filename) {
-  .Call(`_fastMatMR_fmm_to_sparse_Matrix`, filename)
+cpp_fmm_to_sparse_Matrix <- function(filename) {
+  .Call(`_fastMatMR_cpp_fmm_to_sparse_Matrix`, filename)
 }
 
 vec_to_fmm <- function(r_vec, filename) {
