### 2023-10-22 079dd61
DOC: Update the return values for functions
Closes gh-39.
