+++ R/fastMatMR-package.R
@@ -63,12 +63,11 @@ fmm_to_sparse_Matrix <- function(filename) {
 #' @name vec_to_fmm
 #' @title Convert a Numeric Vector to Matrix Market Format
 #' @description This function takes a numeric vector and converts it into a
-#'   Matrix Market file.
+#'   Matrix Market output file.
 #' @param input A numeric vector to be converted.
 #' @param filename The name of the output file where the Matrix Market formatted
 #'   data will be saved.
-#' @return This function has no return value. It writes a Matrix Market
-#'   formatted file to disk.
+#' @return A boolean indicating success or failure. Writes a MTX file to disk.
 #' @examples
 #' \dontrun{
 #' vec <- c(1, 2, 3)
@@ -85,8 +84,7 @@ NULL
 #' @param input A numeric matrix to be converted.
 #' @param filename The name of the output file where the Matrix Market formatted
 #'   data will be saved.
-#' @return This function has no return value. It writes a Matrix Market
-#'   formatted file to disk.
+#' @return A boolean indicating success or failure. Writes a MTX file to disk.
 #' @examples
 #' \dontrun{
 #' mat <- matrix(c(1, 2, 3, 4), nrow = 2)
@@ -103,8 +101,7 @@ NULL
 #' @param input A sparse numeric matrix to be converted.
 #' @param filename The name of the output file where the Matrix Market formatted
 #'   data will be saved.
-#' @return This function has no return value. It writes a Matrix Market
-#'   formatted file to disk.
+#' @return A boolean indicating success or failure. Writes a MTX file to disk.
 #' @examples
 #' \dontrun{
 #' sparse_mat <- Matrix::Matrix(c(1, 0, 0, 2), nrow = 2, sparse = TRUE)
@@ -124,8 +121,7 @@ NULL
 #'   where the Matrix Market formatted data will be saved.  It is recommended to
 #'   use a filename ending with ".mtx" for clarity.
 #'
-#' @return This function has no return value. It writes a Matrix Market
-#'   formatted file to disk.
+#' @return A boolean indicating success or failure. Writes a MTX file to disk.
 #'
 #' @examples
 #' \dontrun{
