### 2023-10-23 eb600a4
ENH: Add support for integer vectors
Closes gh-25

Needs the auto-gen data
