### 2023-10-23 bc92d97
MAINT: Add auto-gen data for integer vector inputs
Part of gh-25
