+++ man/intvec_to_fmm.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/fastMatMR-package.R
-\name{vec_to_fmm}
-\alias{vec_to_fmm}
-\title{Convert a Numeric Vector to Matrix Market Format}
+\name{intvec_to_fmm}
+\alias{intvec_to_fmm}
+\title{Convert a numeric integer vector to Matrix Market Format}
 \arguments{
-\item{input}{A numeric vector to be converted.}
+\item{input}{A numeric integer vector to be converted.}
 
 \item{filename}{The name of the output file where the Matrix Market formatted
 data will be saved.}
@@ -13,12 +13,12 @@ data will be saved.}
 A boolean indicating success or failure. Writes a MTX file to disk.
 }
 \description{
-This function takes a numeric vector and converts it into a
+This function takes a numeric intvector and converts it into a
 Matrix Market output file.
 }
 \examples{
 \dontrun{
-vec <- c(1, 2, 3)
-vec_to_fmm(vec, "vector.mtx")
+intvec <- c(1, 2, 3)
+intvec_to_fmm(intvec, "intvector.mtx")
 }
 }
