### 2023-10-23 024898a
ENH: Add support for integer matrices
Also addresses #25.
