### 2023-10-23 6b0a03b
MAINT: Add autogen for integer matrices
