+++ src/to_file.cpp
@@ -46,10 +46,8 @@ bool intvec_to_fmm(cpp11::integers r_vec, std::string filename) {
 [[cpp11::register]] //
 bool mat_to_fmm(cpp11::doubles_matrix<> r_mat,
                std::string filename) {
-  // Get matrix dimensions
   int nrows = r_mat.nrow();
   int ncols = r_mat.ncol();
-
   std::vector<double> std_vec(nrows * ncols);
   int k = 0;
   for (int i = 0; i < nrows; ++i) {
@@ -57,15 +55,12 @@ bool mat_to_fmm(cpp11::doubles_matrix<> r_mat,
       std_vec[k++] = r_mat(i, j);
     }
   }
-
   fmm::matrix_market_header header(nrows, ncols);
-
   std::filesystem::path file_path(filename);
   std::ofstream os(file_path);
   if (!os.is_open()) {
     return false;
   }
-
   fmm::write_matrix_market_array(os, header, std_vec);
   os.close();
   return true;
@@ -100,13 +95,10 @@ bool sparse_to_fmm(cpp11::sexp input, std::string filename) {
   std::vector<double> x_vec =
       cpp11::as_cpp<std::vector<double>>(input.attr("x"));
   std::vector<int> dim_vec = cpp11::as_cpp<std::vector<int>>(input.attr("Dim"));
-
   fmm::matrix_market_header header(dim_vec[0], dim_vec[1]);
-
   // Check if the matrix is a diagonal matrix by its class
   std::string matrix_class =
       cpp11::as_cpp<std::string>(cpp11::as_sexp(input.attr("class")));
-
   if (matrix_class == "ddiMatrix") {
     std::vector<double> x_vec =
         cpp11::as_cpp<std::vector<double>>(cpp11::as_sexp(input.attr("x")));
@@ -120,15 +112,11 @@ bool sparse_to_fmm(cpp11::sexp input, std::string filename) {
     p_vec = cpp11::as_cpp<std::vector<int>>(cpp11::as_sexp(input.attr("p")));
     i_vec = cpp11::as_cpp<std::vector<int>>(cpp11::as_sexp(input.attr("i")));
   }
-
   std::ofstream os(filename);
   if (!os.is_open()) {
     return false;
   }
-
   fmm::write_matrix_market_csc(os, header, p_vec, i_vec, x_vec, false);
-
   os.close();
-
   return true;
 }
