+++ R/cpp11.R
@@ -24,6 +24,10 @@ mat_to_fmm <- function(r_mat, filename) {
   .Call(`_fastMatMR_mat_to_fmm`, r_mat, filename)
 }
 
+intmat_to_fmm <- function(r_mat, filename) {
+  .Call(`_fastMatMR_intmat_to_fmm`, r_mat, filename)
+}
+
 sparse_to_fmm <- function(input, filename) {
   .Call(`_fastMatMR_sparse_to_fmm`, input, filename)
 }
