### 2023-10-23 7ba2adc
BUG: Silently load Matrix for sparse read ops
Closes gh-22
