### 2023-10-23 HaoZeke/fastMatMR#27 @HaoZeke
Actually users should not be calling this function either. `write_fmm` is meant to be used.
