### 2023-10-23 HaoZeke/fastMatMR#51 @HaoZeke
For C++ functions I'm not sure this makes any sense...
