### 2023-10-23 HaoZeke/fastMatMR#49 @HaoZeke
If `os.open()` fails the function will return `FALSE`, and nothing else needs to be done.
