### 2023-10-23 HaoZeke/fastMatMR#26 @HaoZeke
Users should use `write_fmm` instead..
