### 2023-10-23 566fe20
MAINT: Cleanup and use is_matrix_loaded more
