### 2023-10-23 0a70255
MAINT: Consistent naming
Closes gh-29.
