+++ R/fastMatMR-package.R
@@ -126,9 +126,9 @@ NULL
 #' }
 NULL
 
-#' @export sparse_to_fmm
-#' @rdname sparse_to_fmm
-#' @name sparse_to_fmm
+#' @export sparse_Matrix_to_fmm
+#' @rdname sparse_Matrix_to_fmm
+#' @name sparse_Matrix_to_fmm
 #' @title Convert a Sparse Numeric Matrix to Matrix Market Format
 #' @description This function takes a sparse numeric matrix and converts it into
 #'   a Matrix Market file.
@@ -139,7 +139,7 @@ NULL
 #' @examples
 #' \dontrun{
 #' sparse_mat <- Matrix::Matrix(c(1, 0, 0, 2), nrow = 2, sparse = TRUE)
-#' sparse_to_fmm(sparse_mat, "sparse_matrix.mtx")
+#' sparse_Matrix_to_fmm(sparse_mat, "sparse_matrix.mtx")
 #' }
 NULL
 
@@ -188,7 +188,7 @@ write_fmm <- function(input, filename = "out.mtx") {
       return(mat_to_fmm(input, expanded_fname)) # nolint. C++ function.
     }
   } else if (inherits(input, "sparseMatrix")) {
-    return(sparse_to_fmm(input, expanded_fname)) # nolint. C++ function.
+    return(sparse_Matrix_to_fmm(input, expanded_fname)) # nolint. C++ function.
   } else {
     stop(
       paste(
