+++ man/sparse_Matrix_to_fmm.Rd
@@ -1,7 +1,7 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/fastMatMR-package.R
-\name{sparse_to_fmm}
-\alias{sparse_to_fmm}
+\name{sparse_Matrix_to_fmm}
+\alias{sparse_Matrix_to_fmm}
 \title{Convert a Sparse Numeric Matrix to Matrix Market Format}
 \arguments{
 \item{input}{A sparse numeric matrix to be converted.}
@@ -19,6 +19,6 @@ a Matrix Market file.
 \examples{
 \dontrun{
 sparse_mat <- Matrix::Matrix(c(1, 0, 0, 2), nrow = 2, sparse = TRUE)
-sparse_to_fmm(sparse_mat, "sparse_matrix.mtx")
+sparse_Matrix_to_fmm(sparse_mat, "sparse_matrix.mtx")
 }
 }
