+++ NAMESPACE
@@ -6,7 +6,7 @@ export(fmm_to_vec)
 export(intmat_to_fmm)
 export(intvec_to_fmm)
 export(mat_to_fmm)
-export(sparse_to_fmm)
+export(sparse_Matrix_to_fmm)
 export(vec_to_fmm)
 export(write_fmm)
 useDynLib(fastMatMR, .registration = TRUE)
