### 2023-10-23 HaoZeke/fastMatMR#29 @HaoZeke
Also, there are some minor differences between Matrix Market files and what `fast matrix market` supports, including higher than single precision digits. Since the package name is abbreviated to `fmm` (in C++). `fmm_to_sparse_Matrix` is different because it constructs the sparse matrix as implemented in `Matrix`, a `dgCMatrix`.

`sparse_to_fmm` has been converted to `sparse_Matrix_to_fmm`
