### 2023-10-23 0121c41
TST: Add new tests for integer inputs and outputs
