+++ src/helpers.cpp
@@ -1,5 +1,5 @@
-#include "cpp11.hpp"
 #include "helpers.h"
+#include "cpp11.hpp"
 
 bool is_matrix_loaded() {
   SEXP matrixNamespace = R_FindNamespace(Rf_mkString("Matrix"));
