### 2023-10-24 7aacb15
MAINT: Cleanup vignette and link to intro
Closes gh-41.
