### 2023-10-24 e0646fe
DOC: Fix overritten variables in examples
Closes gh-40.
