### 2023-10-24 HaoZeke/fastMatMR#38 @HaoZeke
Yes. All scales were considered, the plots use a mix of `log-log` and semi-log to show the trends, and all of them are interpreted in the text as well.
