### 2023-10-24 HaoZeke/fastMatMR#42 @HaoZeke
The examples can be run locally, for saving, and then those can be used subsequently for loading. The speed increases are largely by using C++ to load data in an optimal manner so this isn't feasible.
