### 2023-10-24 HaoZeke/fastMatMR#46 @HaoZeke
This is loaded in as part of the `cpp11` package. It is found by `R` by using the following directive in the `DESCRIPTION` (as suggested by the documentation):

```
LinkingTo:
    cpp11
```

There really isn't a good way to get an IDE to be aware of the R installed package headers.
