### 2023-10-24 HaoZeke/fastMatMR#28 @HaoZeke
This is a significant feature request, and can be done without much difficulty outside of `R` at the moment. Will be handled in a later release.
