### 2023-10-24 db15948
BLD: Update pixi to fast_matrix_market 1.7.4
Pixi update
