+++ pixi.toml
@@ -13,7 +13,7 @@ r-renv = "1.0.2.*"
 compilers = "1.6.0.*"
 radian = "0.6.7.*"
 ipython = "8.14.0.*"
-fast_matrix_market = "1.7.2.*"
+fast_matrix_market = "1.7.4.*"
 numpy = "1.25.2.*"
 scipy = "1.11.2.*"
 towncrier = "23.6.0.*"
