### 2023-10-24 2cd5042
Bump version to 1.1.0
