### 2023-10-24 #606 @HaoZeke
Sorry for the delay. Thank you reviewers and editors. I've made changes, in accordance to the suggestions and comments, and made a new release.

# Response to @osorensen
- [X] **Dense matrices with the matrix package**

Response in https://github.com/ropensci/software-review/issues/606#issuecomment-1731692299.

- [X] **Statement of need**
> The README file contains a section titled “Why?”, but it does not state explicitly what problems the software is designed to solved nor what the target audience is

I have rewritten the [readme](https://github.com/HaoZeke/fastMatMR#why) to explain the context.

- [X] **Vignettes**

> There are vignettes, but they are very brief. Please provide some more extended examples, with more text explaining what is being done.

The functionality of the package is to read and write a particular file format, commonly used in data-science across many languages including `R`. Apart from the dense matrix and vector support, the vignettes show performance improvements over the implementation in `Matrix`.

- [X] **Examples**
> I understand that the fmm_to_* functions need an mtx file to work, and I guess it’s therefore the examples with these functions are marked with \dontrun{}. However, wouldn’t it possible to define a character vector in R that contains the required input, and then feed this to fmm_to_* in order to have an example that can actually be run?

The `write_fmm` functions can be run locally, for saving R objects as shown, and then those can be used subsequently for loading. The speed increases are largely by using C++ to load data in an optimal manner so this isn't feasible. https://github.com/HaoZeke/fastMatMR/issues/42

# Response to @czeildi
> All in all, I could easily install the package, understand its functionality
and use it in some of my (simple) own examples without having previous knowledge of matrix market. I could navigate the code and mostly understand it (I am not experienced with cpp11 code).

Glad to hear it :)

> I have a couple of comments, detailed below. Most of them are non-blocking and
might be somewhat opinionated so feel free to apply them as you see fit (besides the ones where I did not tick the corresponding box)

These were all excellent, and I addressed each of them in the linked issues, I will briefly link to them again.

> Performance: I did not re-run the benchmark vignettes myself but tried the package
    out with some moderately large matrices and it felt reasonably fast. 👍🏼

Glad to hear that. The benchmark vignettes are very large and not really meant to be run often (they can take upto 30-40 minutes).

>    it seems the lib cannot handle paths with ~ (?) If this cannot be made to work easily, some error message would be very helpful, I think it would be needed
    to be added here: https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L26 The corresponding read file returns an error message: 'cannot open file' which is more helpful.
    (I am guessing the issue is related to ~ in path because write_fmm(vec, "~/vector.mtx") returned false for me while write_fmm(vec, "/home/ildi_home/vector.mtx") worked)

This was a pretty involved feature, but I managed to get it in https://github.com/HaoZeke/fastMatMR/issues/24.

>   I really appreciated the precise error message in
```
vec_to_fmm(1L, "vector.mtx")
#> Error: Invalid input type, expected 'double' actual 'integer'
```
> However, I was wondering if it could actually accept integers as well?

Implemented in https://github.com/HaoZeke/fastMatMR/issues/25.

>improve error message? It is a matrix, just sparse, not plain
```
mat_to_fmm(sparse_mat, "sparse_matrix.mtx")
#> Error in mat_to_fmm(sparse_mat, "sparse_matrix.mtx") : object is not a matrix
```

As noted in https://github.com/HaoZeke/fastMatMR/issues/26, users should be calling `write_fmm` instead of the lower level functions.

> improve error message?
```
mat <- matrix(c(1, 2, 3, 4), nrow = 2)
sparse_to_fmm(mat, "matrix.mtx")
#> Error: Invalid input type, expected 'double' actual 'NULL'
```

Same as above (https://github.com/HaoZeke/fastMatMR/issues/27), `write_fmm` has the right user-facing semantics.

> future suggestion: support reading .mtx.gz files as I can download from for example https://math.nist.gov/MatrixMarket/data/NEP/quebec/qh1484.html ?

This one is tracked, but not for this release. It would need quite a bit more design and I'm not sure it will be in scope. Tracked here: https://github.com/HaoZeke/fastMatMR/issues/28

> consider increased consistency of function names? e.g fmm_to_sparseMatrix vs sparse_to_fmm. sparse_to_fmm was a bit confusing to me in that if I understand correctly, the package is 'fast', not the target format itself (?). matrix and vector is abbreviated to spare a few characters, while sparse_matrix is not. A vague suggestion: possibly follow the pattern of fmm_write_vector, fmm_read_to_vector for all formats?

As noted in https://github.com/HaoZeke/fastMatMR/issues/29,
sparse_to_fmm has been converted to sparse_Matrix_to_fmm for consistency.

**Documentation**

>    I did not tick function documentation because although they exist, argument names
    (fname vs filename) are not always consistent between the documentation and the code
    itself

I have fixed this in https://github.com/HaoZeke/fastMatMR/issues/30.

>    I did not tick Examples, because although the examples run, but only in the right order, i.e. if I first run the examples of the write
    functions then the examples of the read functions run correctly but they are not self contained

Completed in https://github.com/HaoZeke/fastMatMR/issues/42 and https://github.com/HaoZeke/fastMatMR/issues/31. Essentially, while the `write_fmm` can be self-contained, `fmm_to_*` need to have matrix market files present.

>    readme: "Unlike the Matrix package" : maybe rephrase? it is a bit off-putting for me to start the whole readme with a negative comparison, although
    I understand that a core part of the motivation for this package is that Matrix does not support everything and not fast enough. Even changing the order of the two parts of the sentence would feel better for me but this is really personal
>    In the readme: "we support / the package supports". Consider unifying these two wording styles

Completed in https://github.com/HaoZeke/fastMatMR/issues/32 and https://github.com/HaoZeke/fastMatMR/issues/33. Along with the suggestions from the other reviewer, the `readme` has been more or less completely re-written, and I hope it is clearer now.

>    rename basic_usage to getting started? The vignette title is getting started but you can see basic_usage in the url

Completed in https://github.com/HaoZeke/fastMatMR/issues/34.

>    clarify what the addendum is about in the getting started readme? If I understand correctly, it demonstrates the incorrect handling of NA by Matrix package?

Completed in https://github.com/HaoZeke/fastMatMR/issues/35.

>    Maybe add a bit more context on why would I want to read back in Python? Possibly rather focus on being a language agnostic format?

Completed in https://github.com/HaoZeke/fastMatMR/issues/36.

>    consider moving the vignette summary to the beginning/remove it? It felt to me like repeating basically the same information as the beginning of the vignette

Completed in https://github.com/HaoZeke/fastMatMR/issues/37.

>    Did you try out logarithmic scale for the benchmark plots? Would they possibly be more readable?

Yes. All scales were considered, the plots use a mix of log-log and semi-log to show the trends, and all of them are interpreted in the text as well. Completed in https://github.com/HaoZeke/fastMatMR/issues/38.

>    In several function documentation you write "This function has no return value." This is technically not true, all functions return NULL if they have no other return value
    but rather these functions are not called for their return value, but for their side effect? Maybe see how other packages handle write functions?

Completed in https://github.com/HaoZeke/fastMatMR/issues/39.

>    write_fmm example: sparse_mat variable is overwritten before it is written to disk so it is not easy to actually run both of the two variants included

Completed in https://github.com/HaoZeke/fastMatMR/issues/40.

>    readme sentence: "Similarly, we support writing and reading from other R objects (e.g. standard R vectors and matrices), as seen in the basic use vignette." Consider rephrasing a bit, e.g. refer to 'getting started' vignette instead of basic use, possibly as described in the vignette / as you can see in the vignette instead of as seen?

Rewritten slightly for the new readme.

>    I am not sure how, but I think there is a relatively easy way to have the getting started vignette show up as a separate menu item in the packagedown website and not just among articles, that could improve its discoverability

Completed in https://github.com/HaoZeke/fastMatMR/issues/41.

**Code**

>    I see that R CMD check seem to give a note about pixi files, it seems pixi.* might not
    be the correct syntax in .Rbuildignore? (I do not know what the correct syntax would be,
    but the check gives a note both in github actions and locally it seems. If all else fails, adding the files one by one instead of a wildcard syntax definitely works)
    package version in DESCRIPTION does not match package version in news, is this intended?

Completed in https://github.com/HaoZeke/fastMatMR/issues/43.

>    Ideally, tests would not write to the project root in my opinion even if those files are added
    to .gitignore, but rather write to a temp directory. https://testthat.r-lib.org/articles/test-fixtures.html might help find a more robust solution.
    the readability of read unit tests could be improved if all tests are self contained and independent of each other: e.g. you can run any test_that block in itself and it is a working test (no setup code before test_that blocks ideally)

Completed in https://github.com/HaoZeke/fastMatMR/issues/44.

>    I think there is no need to add library(testthat) to test files

Completed in https://github.com/HaoZeke/fastMatMR/issues/45.

>    consistency of indentation styles could be a bit improved, consider running styler::style_pkg, see for example https://github.com/HaoZeke/fastMatMR/blob/main/tests/testthat/test-write_fmm.R#L34

These are run already, but the tests were refactored so I think it should be OK now. Completed in https://github.com/HaoZeke/fastMatMR/issues/45.

>    this is really overly pedantic, but you could consider specifying which rule to ignore in the #nolint comments instead of ignoring every rule (it is considered generally a best practice, but does not have real impact here). This is not needed if you would need to ignore several rules

Since these are C++ functions, the linting is basically for everything. Completed in https://github.com/HaoZeke/fastMatMR/issues/51.

>    pre commit github workflow name is a bit misleading to me, I understand you could add the same functionality as a pre commit hook, but a github action is a post commit/push thing, maybe title it after what it does, e.g. style checking?

Completed in https://github.com/HaoZeke/fastMatMR/issues/47.

>    where is this included from? I only see cpp11.cpp https://github.com/HaoZeke/fastMatMR/blob/main/src/from_file.cpp#L10
    My IDE underlined this with red, but it might be fine, I am not familiar with cpp bindings in R packages

This is loaded in as part of the `cpp11` package. It is found by `R` by using the following directive in the `DESCRIPTION` (as suggested by the documentation):

```
LinkingTo:
    cpp11
```

There really isn't a good way to get an IDE to be aware of the R installed package headers. `cpp11.cpp` is actually auto-generated :)

Completed in https://github.com/HaoZeke/fastMatMR/issues/46.

>    commented out code should be removed (https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L20)

Done in https://github.com/HaoZeke/fastMatMR/issues/48.

>    what happens if cpp lib call fails for some reason? Will the file handle be closed properly? See https://github.com/HaoZeke/fastMatMR/blob/main/src/to_file.cpp#L30

If `os.open()` fails the function will return `FALSE`, and nothing else needs to be done, completed in https://github.com/HaoZeke/fastMatMR/issues/49.

>    possibly move rebuild-benchmarks to a tools sub directory or similar? https://r-pkgs.org/misc.html#sec-misc-tools.

Completed in https://github.com/HaoZeke/fastMatMR/issues/50.

============================

Thank you all for the great comments!
