### 2023-10-25 #606 @czeildi
@HaoZeke great work and thank you for your detailed response and making your changes easy to track. I am now happy to recommend approving this package :)

I still have two small comments (neither blocking approval):
- I think there is a typo in .Rbuildignore, I submitted a PR to fix: https://github.com/HaoZeke/fastMatMR/pull/63

> As noted in https://github.com/HaoZeke/fastMatMR/issues/26, users should be calling write_fmm instead of the lower level functions.

If the users are not supposed to be calling the lower level functions, do they need to be exported, and thus being part of the public user interface of the package? It could be less confusing possibly if only those functions are exported which are meant to be called directly by users.
