### 2023-10-25 #606 @HaoZeke
> @HaoZeke great work and thank you for your detailed response and making your changes easy to track. I am now happy to recommend approving this package :)

Wonderful, thanks so much ^_^

>
> I still have two small comments (neither blocking approval):
>
> * I think there is a typo in .Rbuildignore, I submitted a PR to fix: [Fix typo in  .Rbuildignore HaoZeke/fastMatMR#63](https://github.com/HaoZeke/fastMatMR/pull/63)

Merging soon :)

> > As noted in [HaoZeke/fastMatMR#26](https://github.com/HaoZeke/fastMatMR/issues/26), users should be calling write_fmm instead of the lower level functions.
>
> If the users are not supposed to be calling the lower level functions, do they need to be exported, and thus being part of the public user interface of the package? It could be less confusing possibly if only those functions are exported which are meant to be called directly by users.

I thought about this, but decided not to mainly because downstream package users (including me for other dependent packages) would want to call these functions where it makes sense (e.g. if my downstream package will only write matrix data or sparse matrix data) and [CRAN disallows](https://community.rstudio.com/t/how-to-include-internal-functions/38654/8) / warns [calling unexported functions](https://blog.r-hub.io/2019/12/12/internal-functions/) via `:::`
