### 2023-10-25 HaoZeke/fastMatMR#64 @HaoZeke
MAINT: Possibly hide functions which shouldn't be called by users
> If the users are not supposed to be calling the lower level functions, do they need to be exported, and thus being part of the public user interface of the package? It could be less confusing possibly if only those functions are exported which are meant to be called directly by users.

First noted here: https://github.com/ropensci/software-review/issues/606#issuecomment-1779696507
