### 2023-10-25 HaoZeke/fastMatMR#64 @HaoZeke
Response from here: https://github.com/ropensci/software-review/issues/606#issuecomment-1779899582

I thought about this, but decided not to mainly because downstream package users (including me for other dependent packages) would want to call these functions where it makes sense (e.g. if my downstream package will only write matrix data or sparse matrix data) and [CRAN disallows](https://community.rstudio.com/t/how-to-include-internal-functions/38654/8) / warns [calling unexported functions](https://blog.r-hub.io/2019/12/12/internal-functions/) via :::
