### 2023-10-30 #606 @ropensci-review-bot
Approved! Thanks @HaoZeke for submitting and @osorensen, @czeildi for your reviews! :grin:
