### 2023-10-30 #606 @HaoZeke
Sorry to bring this up here, but, I can't seem to get the centralized documentation and r-universe builds to run, @maelle would you have any suggestions?

- I moved the repo
- Got back admin access
- Pushed a few commits changing links
  - I thought at this stage a build would run on r-universe based on [the documentation](https://devguide.ropensci.org/ci.html#even-more-ci-ropensci-docs)
  - It doesn't seem to have a job defined (https://dev.ropensci.org/fastMatMR)
- Removed my customized `pkgdown` stuff
  - This also didn't seem to trigger a build from the centralized documentation server, and the link which should go to the package hits a 404 (https://docs.ropensci.org/fastMatMR)

P.S. Thanks all for the help, this was a really useful process :)
