### 2023-10-30 #606 @HaoZeke
Also perhaps @mpadge would be better suited to offer advice on https://github.com/ropensci/software-review/issues/606#issuecomment-1785287024?
