### 2023-10-30 #606 @mpadge
@HaoZeke r-universe only builds at most every few hours, so my advice would be just wait and build should appear
