### 2023-10-30 #606 @maelle
https://docs.ropensci.org/fastMatMR/ is up :tada:

The dev version of the dev guide has a package maintainer cheatsheet: https://devdevguide.netlify.app/maintenance_cheatsheet

Thanks so much @HaoZeke for your participation in the process!
