### 2023-11-02 #606 @HaoZeke
Now [on CRAN](https://cran.r-project.org/web/packages/fastMatMR/index.html)!

![image](https://github.com/alugowski/fast_matrix_market/assets/4336207/e88624f8-641c-4f70-8f52-3ca636b6c7cd)
